+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.2.0 (unreleased)
 
+Feature:
+  - Dialog to add guides to the document after placing the textures
+
 Fix:
   - fix offsets when placing higher zoom level textures than documents base zoom level
   - additional sanity check when parsing info from document name